% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.block}
\alias{add.block}
\title{Add a new block to the block structure in a recording, automatically incrementing the block id.}
\usage{
add.block(recording, block)
}
\arguments{
\item{recording}{A recording.}

\item{block}{A block structure.}
}
\value{
The recording with the given block added to the block structure.
}
\description{
Add a new block to the block structure in a recording, automatically incrementing the block id.
}
\seealso{
Other recording: \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

