% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.segment.timestamp}
\alias{add.segment.timestamp}
\title{Add timestamps to the segments in the results collected from a recording}
\usage{
add.segment.timestamp(recording, results)
}
\arguments{
\item{recording}{A recording.}

\item{results}{The results collected from a recording as a data frame.}
}
\value{
The results with the segment timestamps added.
}
\description{
Add timestamps to the segments in the results collected from a recording
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

