% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\name{add_segment_timestamp}
\alias{add_segment_timestamp}
\title{[deprecated] Add timestamps to the segments in the results collected from a recording.
Timestamp value gives the position of the _midpoint_ of the respective segment.}
\usage{
add_segment_timestamp(recording, results)
}
\arguments{
\item{recording}{A recording.}

\item{results}{The results collected from a recording as a data frame.}
}
\value{
The results with the segment timestamps added.
}
\description{
[deprecated] Add timestamps to the segments in the results collected from a recording.
Timestamp value gives the position of the _midpoint_ of the respective segment.
}
\seealso{
Other recording: \code{\link{add_block}},
  \code{\link{block_to_seconds}},
  \code{\link{collect_results_collection}},
  \code{\link{collect_results}},
  \code{\link{create_block_simple}},
  \code{\link{create_block_structure}},
  \code{\link{cut_recording_s}},
  \code{\link{cut_recordings_s}},
  \code{\link{cut_recordings}},
  \code{\link{cut_recording}},
  \code{\link{find_recording_overlap}},
  \code{\link{load_recording}},
  \code{\link{new_recording}},
  \code{\link{read.block.event.data}},
  \code{\link{read.config.excel}},
  \code{\link{read.config.sqlite}},
  \code{\link{read.config}},
  \code{\link{read.dbtable.excel}},
  \code{\link{read.dbtable.sqlite}},
  \code{\link{recording_get_zerotime}},
  \code{\link{recording_set_blocks}},
  \code{\link{recording_set_events}},
  \code{\link{recording_set_ibi}},
  \code{\link{recording_set_zerotime}},
  \code{\link{save_recording}}
}

