% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrs_detect.R
\name{adjust_indices}
\alias{adjust_indices}
\title{Adjust two vectors with indices so that these form
pairs of indices, where for each pairthe index in the first
vector always is smaller than the index in the second vector,}
\usage{
adjust_indices(ind.start, ind.stop)
}
\arguments{
\item{ind.start}{start index}

\item{ind.stop}{stop index}
}
\value{
A list with the start and stop indices in the.
}
\description{
Adjust two vectors with indices so that these form
pairs of indices, where for each pairthe index in the first
vector always is smaller than the index in the second vector,
}

