% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyse.frequencydomain.helper}
\alias{analyse.frequencydomain.helper}
\title{Helper function for frequency-domain HRV analysis.}
\usage{
analyse.frequencydomain.helper(metric, spec, settings)
}
\arguments{
\item{metric}{A frequency-domain metric to be estimated.}

\item{spec}{The Lomb-Scargle spectrum estimated from the IBI series using the function lombscargle.}

\item{settings}{The settings structure used for HRV analysis.}
}
\value{
The value of the metric.
}
\description{
Helper function for frequency-domain HRV analysis.
}
\seealso{
Other HRV frequency-domain: \code{\link{analyse.frequencydomain}};
  \code{\link{ibi.band.custom}};
  \code{\link{ibi.band.standard}};
  \code{\link{lombscargle}}
}
\keyword{internal}

