% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyse.frequencydomain}
\alias{analyse.frequencydomain}
\title{Frequency-domain HRV analysis.}
\usage{
analyse.frequencydomain(metric.list, ibi, t.ibi, settings)
}
\arguments{
\item{metric.list}{A list of frequency-domain metrics as strings.}

\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{t.ibi}{A vector with the times of occurrence of the interbeat intervals (IBIs).}

\item{settings}{The settings structure used for HRV analysis with the following fields:

\describe{
\item{$frequencydomain$parameters$demean}{Should the mean be subtraced from the spectrum. Default is TRUE.}
\item{$frequencydomain$parameters$smooth}{Should the spectrum be smoothed. Default is FALSE.}
\item{$frequencydomain$parameters$smooth.kernel}{The smoothing kernel. The default is to use a 'daniell' kernel.}
\item{$frequencydomain$parameters$smooth.degree}{The degree of smoothing. Default is 5.}
\item{$frequencydomain$parameters$normalization}{How to normalize the spectrum. Possible values are 'variance' (default) and 'none'.}
}}
}
\value{
A list with the values of the metrics.
}
\description{
Frequency-domain HRV analysis.
}
\seealso{
Other HRV frequency-domain: \code{\link{analyse.frequencydomain.helper}};
  \code{\link{ibi.band.custom}};
  \code{\link{ibi.band.standard}};
  \code{\link{lombscargle}}
}

