% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyse.geometric.helper}
\alias{analyse.geometric.helper}
\title{Helper function for geometric HRV analysis.}
\usage{
analyse.geometric.helper(metric, x, settings)
}
\arguments{
\item{metric}{A metric to be estimated.}

\item{settings}{The settings structure used for HRV analysis.}

\item{spec}{The Lomb-Scargle spectrum estimated from the IBI series using the function lombscargle.}
}
\value{
The value of the metric.
}
\description{
Helper function for geometric HRV analysis.
}
\seealso{
Other HRV geometric: \code{\link{analyse.geometric}};
  \code{\link{ibi.sd1}}; \code{\link{ibi.sd2}};
  \code{\link{ibi.tinn}};
  \code{\link{ibi.triangular.index}}
}
\keyword{internal}

