% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyse.geometric}
\alias{analyse.geometric}
\title{Geometric HRV analysis.}
\usage{
analyse.geometric(metric.list, ibi, t.ibi = NULL, settings)
}
\arguments{
\item{metric.list}{A list of geometric metrics as strings.}

\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{t.ibi}{A vector with the times of occurrence of the interbeat intervals (IBIs).}

\item{settings}{The settings structure used for HRV analysis with the following fields:

\describe{
\item{$geometric$triangular.index$bindwidth}{The bin width used in the calculation of the triangular index.}
}}
}
\value{
A list with the values of the metrics.
}
\description{
Geometric HRV analysis.
}
\seealso{
Other HRV geometric: \code{\link{analyse.geometric.helper}};
  \code{\link{ibi.sd1}}; \code{\link{ibi.sd2}};
  \code{\link{ibi.tinn}};
  \code{\link{ibi.triangular.index}}
}

