% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyse.nonlinear.helper}
\alias{analyse.nonlinear.helper}
\title{Helper function for nonlinear HRV analysis.}
\usage{
analyse.nonlinear.helper(metric, x, settings)
}
\arguments{
\item{metric}{A frequency-domain metric to be estimated.}

\item{x}{A vector with the interbeat intervals (IBIs).}

\item{settings}{The settings structure used for HRV analysis.}
}
\value{
The value of the metric.
}
\description{
Helper function for nonlinear HRV analysis.
}
\seealso{
Other HRV nonlinear: \code{\link{analyse.nonlinear}};
  \code{\link{ibi.entropy}}
}
\keyword{internal}

