% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyse.timedomain.helper}
\alias{analyse.timedomain.helper}
\title{Helper function for time-domain HRV analysis.}
\usage{
analyse.timedomain.helper(metric, ibi, settings)
}
\arguments{
\item{metric}{A time-domain metric to be estimated.}

\item{ibi}{An array with interbeat intervals.}

\item{settings}{The settings structure used for HRV analysis.}
}
\value{
The value of the metric.
}
\description{
Helper function for time-domain HRV analysis.
}
\keyword{internal}

