% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyse.timedomain}
\alias{analyse.timedomain}
\title{Time-domain HRV analysis.}
\usage{
analyse.timedomain(metric.list, settings, ibi, t.ibi = NULL)
}
\arguments{
\item{metric.list}{A list of time-domain metrics as strings.}

\item{settings}{The settings structure used for HRV analysis (see \code{\link{settings.template}}).}

\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{t.ibi}{A vector with the times of occurrence of the interbeat intervals (IBIs).}
}
\value{
A list with the values of the metrics.
}
\description{
Time-domain HRV analysis.
}
\seealso{
Other HRV time-domain: \code{\link{ibi.mean}};
  \code{\link{ibi.nnx}}; \code{\link{ibi.pnnx.helper}};
  \code{\link{ibi.pnnx}}; \code{\link{ibi.rmssd}};
  \code{\link{ibi.stdev}}; \code{\link{ibi.var}}
}

