% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hrv_frequencydomain.R
\name{analyse_frequencydomain_helper}
\alias{analyse_frequencydomain_helper}
\title{Helper function for frequency-domain HRV analysis.}
\usage{
analyse_frequencydomain_helper(metric, spec, settings)
}
\arguments{
\item{metric}{A frequency-domain metric to be estimated.}

\item{spec}{The Lomb-Scargle spectrum estimated from the IBI series using the function lombscargle.}

\item{settings}{The settings structure used for HRV analysis.}
}
\value{
The value of the metric.
}
\description{
Helper function for frequency-domain HRV analysis.
}
\seealso{
Other HRV frequency-domain: \code{\link{analyse_frequencydomain}},
  \code{\link{ibi_band_custom}},
  \code{\link{ibi_band_standard}},
  \code{\link{lombscargle}}
}
\keyword{internal}

