% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis_hrv_geometric.R
\name{analyse_geometric}
\alias{analyse_geometric}
\title{Geometric HRV analysis.}
\usage{
analyse_geometric(metric.list, ibi, t.ibi = NULL, settings)
}
\arguments{
\item{metric.list}{A list of geometric metrics as strings.}

\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{t.ibi}{A vector with the times of occurrence of the interbeat intervals (IBIs).}

\item{settings}{The settings structure used for HRV analysis with the following fields:

\describe{
\item{$geometric$triangular.index$bindwidth}{The bin width used in the calculation of the triangular index.}
}}
}
\value{
A list with the values of the metrics.
}
\description{
Geometric HRV analysis.
}
\seealso{
Other HRV geometric: \code{\link{analyse_geometric_helper}};
  \code{\link{ibi_sd1}}; \code{\link{ibi_sd2}};
  \code{\link{ibi_tinn}};
  \code{\link{ibi_triangular_index}}
}

