% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis_hrv_geometric.R
\name{analyse_geometric_helper}
\alias{analyse_geometric_helper}
\title{Helper function for geometric HRV analysis.}
\usage{
analyse_geometric_helper(metric, x, settings)
}
\arguments{
\item{metric}{A metric to be estimated.}

\item{settings}{The settings structure used for HRV analysis.}

\item{spec}{The Lomb-Scargle spectrum estimated from the IBI series using the function lombscargle.}
}
\value{
The value of the metric.
}
\description{
Helper function for geometric HRV analysis.
}
\seealso{
Other HRV geometric: \code{\link{analyse_geometric}};
  \code{\link{ibi_sd1}}; \code{\link{ibi_sd2}};
  \code{\link{ibi_tinn}};
  \code{\link{ibi_triangular_index}}
}
\keyword{internal}

