% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyse_nonlinear}
\alias{analyse_nonlinear}
\title{Nonelinear HRV analysis.}
\usage{
analyse_nonlinear(metric.list, ibi, t.ibi = NULL, settings)
}
\arguments{
\item{metric.list}{A list of frequency-domain metrics as strings.}

\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{t.ibi}{A vector with the times of occurrence of the interbeat intervals (IBIs).}

\item{settings}{The settings structure used for HRV analysis with the following fields:

\describe{
\item{$nonlinear$embedding.dimension}{The embedding dimension used in the calculation of entropy}
}}
}
\value{
A list with the values of the metrics.
}
\description{
Nonelinear HRV analysis.
}
\seealso{
Other HRV nonlinear: \code{\link{analyse_nonlinear_helper}};
  \code{\link{ibi_entropy}}
}

