% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hrv_nonlinear.R
\name{analyse_nonlinear_helper}
\alias{analyse_nonlinear_helper}
\title{Helper function for nonlinear HRV analysis.}
\usage{
analyse_nonlinear_helper(metric, x, settings)
}
\arguments{
\item{metric}{A frequency-domain metric to be estimated.}

\item{x}{A vector with the interbeat intervals (IBIs).}

\item{settings}{The settings structure used for HRV analysis.}
}
\value{
The value of the metric.
}
\description{
Helper function for nonlinear HRV analysis.
}
\seealso{
Other HRV nonlinear: \code{\link{analyse_nonlinear}},
  \code{\link{ibi_entropy}}
}
\keyword{internal}

