% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analysis_pipeline_ibi}
\alias{analysis_pipeline_ibi}
\title{Calculate results.}
\usage{
analysis_pipeline_ibi(settings, data)
}
\arguments{
\item{settings}{The settings structure used for HRV analysis (see \code{\link{settings.template}}).}

\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{t.ibi}{A vector with the times of occurrence of the interbeat intervals (IBIs).}
}
\value{
The analysis results as a named list.
}
\description{
This function defines the analysis pipeline for HRV analysis. If new analysis
functions are incorporated these analysis steps should be added to
this function as well.
}
\seealso{
Other pipeline: \code{\link{analyze_block}};
  \code{\link{analyze_recording_collection}};
  \code{\link{analyze_recording}}
}

