% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyze.block}
\alias{analyze.block}
\title{Analyze results in a block.}
\usage{
analyze.block(recording, settings, block, analysis.pipeline.function)
}
\arguments{
\item{recording}{A recording structure}

\item{settings}{The settings structure used for HRV analysis (see \code{\link{settings.template}}).}

\item{block}{A block structure.}

\item{analysis.pipeline.function}{The pipeline function used in the
analysis, i.e., a funtion containing the individual analysis steps.}
}
\value{
The recording structure with the result from the block added to the results structure in the recording.
}
\description{
Analyze the results in a particular block.
}
\seealso{
Other pipeline: \code{\link{analysis.pipeline.ibi}};
  \code{\link{analyze.recording.collection}};
  \code{\link{analyze.recording}}
}

