% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyze.recording.collection}
\alias{analyze.recording.collection}
\title{Analyze all recordings in a recording collection.}
\usage{
analyze.recording.collection(recording.collection, settings,
  analysis.pipeline.function = analysis.pipeline)
}
\arguments{
\item{settings}{The settings structure used for HRV analysis (see \code{\link{settings.template}}).}

\item{analysis.pipeline.function}{The pipeline function used in the
analysis, i.e., a funtion containing the individual analysis steps.}

\item{recording}{A recording structure}
}
\value{
The recording collection with the result added to each of the individual recordings.
}
\description{
All blocks in all recordings in the recording collection  are analyzed.
}
\seealso{
Other pipeline: \code{\link{analysis.pipeline.ibi}};
  \code{\link{analyze.block}};
  \code{\link{analyze.recording}}
}

