% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyze.recording}
\alias{analyze.recording}
\title{Analyze a recording.}
\usage{
analyze.recording(recording, settings,
  analysis.pipeline.function = analysis.pipeline)
}
\arguments{
\item{recording}{A recording structure}

\item{settings}{The settings structure used for HRV analysis (see \code{\link{settings.template}}).}

\item{analysis.pipeline.function}{The pipeline function used in the
analysis, i.e., a funtion containing the individual analysis steps.}
}
\value{
The recording structure with the result from the individual blocks added to the results structure in the recording.
}
\description{
All blocks in the recording are analyzed.
}
\seealso{
Other pipeline: \code{\link{analysis.pipeline.ibi}};
  \code{\link{analyze.block}};
  \code{\link{analyze.recording.collection}}
}

