% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pipeline.R
\name{analyze_block}
\alias{analyze_block}
\title{Analyze results in a block.}
\usage{
analyze_block(recording, settings, signal, block, analysis.pipeline.function)
}
\arguments{
\item{recording}{A recording structure}

\item{settings}{The settings structure used for the analysis (see e.g., \code{\link{settings_template_hrv}}).}

\item{signal}{A string or a list with the names of the signals to analyze.}

\item{block}{A block structure.}

\item{analysis.pipeline.function}{The pipeline function used in the
analysis, i.e., a funtion containing the individual analysis steps.}

\item{signal}{The name}
}
\value{
The recording structure with the result from the block added to the results structure in the recording.
}
\description{
Analyze the results in a particular block.
}
\seealso{
Other pipeline: \code{\link{analysis_pipeline_ibi}},
  \code{\link{analyze_recording_collection}},
  \code{\link{analyze_recording}}
}

