% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pipeline.R
\name{analyze_recording}
\alias{analyze_recording}
\title{Analyze a recording.}
\usage{
analyze_recording(recording, settings, signal, analysis.pipeline.function)
}
\arguments{
\item{recording}{A recording structure}

\item{settings}{The settings structure used for HRV analysis (see \code{\link{settings.template}}).}

\item{signal}{A string or a list with the names of the signals to analyze.}

\item{analysis.pipeline.function}{The pipeline function used in the
analysis, i.e., a funtion containing the individual analysis steps.}
}
\value{
The recording structure with the result from the individual blocks added to the results structure in the recording.
}
\description{
All blocks in the recording are analyzed.
}
\seealso{
Other pipeline: \code{\link{analysis_pipeline_ibi}},
  \code{\link{analyze_block}},
  \code{\link{analyze_recording_collection}}
}

