% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis_pipeline.R
\name{analyze_recording_collection}
\alias{analyze_recording_collection}
\title{Analyze all recordings in a recording collection.}
\usage{
analyze_recording_collection(recording.collection, settings, signal,
  analysis.pipeline.function)
}
\arguments{
\item{settings}{The settings structure used for HRV analysis (see \code{\link{settings.template}}).}

\item{signal}{A string or a list with the names of the signals to analyze.}

\item{analysis.pipeline.function}{The pipeline function used in the
analysis, i.e., a funtion containing the individual analysis steps.}

\item{recording}{A recording structure}
}
\value{
The recording collection with the result added to each of the individual recordings.
}
\description{
All blocks in all recordings in the recording collection  are analyzed.
}
\seealso{
Other pipeline: \code{\link{analysis_pipeline_ibi}};
  \code{\link{analyze_block}};
  \code{\link{analyze_recording}}
}

