% Generated by roxygen2 (4.0.2): do not edit by hand
\name{block.to.seconds}
\alias{block.to.seconds}
\title{Return a block structure where any timestamps have been converted
to seconds with respect to the start time of the recording.}
\usage{
block.to.seconds(recording, block = NULL,
  timeformat = "\%Y\%m\%dT\%H\%M\%S")
}
\arguments{
\item{recording}{A recording.}

\item{block}{A block structure. If not given, it is read from the field \code{recording$conf$blocks}.}

\item{timeformat}{The time format to use. The default is "\%Y\%m\%dT\%H\%M\%S".}
}
\value{
The block structure with the timestamps replaced by seconds
from the beginning of the recording.
}
\description{
Return a block structure where any timestamps have been converted
to seconds with respect to the start time of the recording.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

