% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blocks.R
\name{block_to_seconds}
\alias{block_to_seconds}
\title{Return a block structure where any timestamps have been converted
to seconds with respect to the start time of the recording.}
\usage{
block_to_seconds(recording, block = NULL,
  timeformat = "\%Y\%m\%dT\%H\%M\%S")
}
\arguments{
\item{recording}{A recording.}

\item{block}{A block structure. If not given, it is read from the field \code{recording$conf$blocks}.}

\item{timeformat}{The time format to use. The default is "\%Y\%m\%dT\%H\%M\%S".}
}
\value{
The block structure with the timestamps replaced by seconds
from the beginning of the recording.
}
\description{
Return a block structure where any timestamps have been converted
to seconds with respect to the start time of the recording.
}
\seealso{
Other recording: \code{\link{add_block}},
  \code{\link{add_segment_timestamp}},
  \code{\link{collect_results_collection}},
  \code{\link{collect_results}},
  \code{\link{create_block_simple}},
  \code{\link{create_block_structure}},
  \code{\link{cut_recording_s}},
  \code{\link{cut_recordings_s}},
  \code{\link{cut_recordings}},
  \code{\link{cut_recording}},
  \code{\link{find_recording_overlap}},
  \code{\link{load_recording}},
  \code{\link{new_recording}},
  \code{\link{read.block.event.data}},
  \code{\link{read.config.excel}},
  \code{\link{read.config.sqlite}},
  \code{\link{read.config}},
  \code{\link{read.dbtable.excel}},
  \code{\link{read.dbtable.sqlite}},
  \code{\link{recording_get_zerotime}},
  \code{\link{recording_set_blocks}},
  \code{\link{recording_set_events}},
  \code{\link{recording_set_ibi}},
  \code{\link{recording_set_zerotime}},
  \code{\link{save_recording}}
}

