% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check.rr.detection}
\alias{check.rr.detection}
\title{Check the quality of R-peak detection.}
\usage{
check.rr.detection(recording, blockid, ecg.signal.name = "ECG",
  ibi.signal.name = "ibi.amp", type = "seconds", interval = 10,
  new.plot = TRUE, filename = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{blockid}{The ID of the measurement block.}

\item{ecg.signal.name}{The name of the ECG signal in the recording structure (default is 'ECG')}

\item{ibi.signal.name}{The name of the interbeat interval (IBI) signal in the recording structure (default is 'ibi.amp', to ensure tha the R-peaks are plotted at the correct height.)}

\item{type}{The type of the x-axis (optional, default is seconds):
\describe{
\item{seconds}{Plot seconds on the x-axis.}
\item{timestamp}{Plot timestamps on the x-axis.}
}}

\item{interval}{The interval at which to plot x-axis markers. Optional, default is 15.}

\item{new.plot}{Boolean indicating whether a new plot window should be opened (optional, default is TRUE).}

\item{filename}{The full path and filename in which to save the figure. If NULl (default), the figure is not saved. The format of the figure (PDF or PNG) is automatically determined from the filename extension. The size of the figure is an A4 in landscape mode.}
}
\value{
Nothing. Produces a plot or saves the figure.
}
\description{
Plot the raw ECG data together with interbeat-interval data. This allows the quality of the R-peak detection to be checked visually.
}
\seealso{
Other visualizations: \code{\link{plot.all.blocks}};
  \code{\link{plot.block}}; \code{\link{plot.ecg.r.peak}};
  \code{\link{plot.metric}}
}

