% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colibri.R
\docType{package}
\name{colibri}
\alias{colibri}
\alias{colibri-package}
\title{Platform for automating analyses of physiological time series in R.}
\description{
The colibri package provides functions for automating analysis of
physiological time series in R.
}

