% Generated by roxygen2 (4.0.2): do not edit by hand
\name{collect.results}
\alias{collect.results}
\title{Collect results.}
\usage{
collect.results(recording, format = "data.frame")
}
\arguments{
\item{recording}{A recording.}

\item{format.}{Output formmat. Either \code{data.frame} (default) or \code{matrix}.}
}
\value{
The analysis results either as a data frame or as a matrix.
}
\description{
Get the analysis results, which are internally stored as a list of matrices,
from a recording and return them as a data frame or as a matrix.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

