% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\name{collect_results}
\alias{collect_results}
\title{Collect results.}
\usage{
collect_results(recording, format = "data.frame", add_timestamp = TRUE)
}
\arguments{
\item{recording}{A recording.}

\item{format}{Output format. Either \code{data.frame} (default) or \code{matrix}.}

\item{add_timestamp}{Should segment timestamps be added. Only works if the output format is a data frame. Boolean. Default is \code{TRUE}.}
}
\value{
The analysis results either as a data frame or as a matrix.
}
\description{
Get the analysis results, which are internally stored as a list of matrices,
from a recording and return them as a data frame or as a matrix.
}
\seealso{
Other recording: \code{\link{add_block}},
  \code{\link{add_segment_timestamp}},
  \code{\link{block_to_seconds}},
  \code{\link{collect_results_collection}},
  \code{\link{create_block_simple}},
  \code{\link{create_block_structure}},
  \code{\link{cut_recording_s}},
  \code{\link{cut_recordings_s}},
  \code{\link{cut_recordings}},
  \code{\link{cut_recording}},
  \code{\link{find_recording_overlap}},
  \code{\link{load_recording}},
  \code{\link{new_recording}},
  \code{\link{read.block.event.data}},
  \code{\link{read.config.excel}},
  \code{\link{read.config.sqlite}},
  \code{\link{read.config}},
  \code{\link{read.dbtable.excel}},
  \code{\link{read.dbtable.sqlite}},
  \code{\link{recording_get_zerotime}},
  \code{\link{recording_set_blocks}},
  \code{\link{recording_set_events}},
  \code{\link{recording_set_ibi}},
  \code{\link{recording_set_zerotime}},
  \code{\link{save_recording}}
}

