% Generated by roxygen2 (4.0.2): do not edit by hand
\name{collect_results_collection}
\alias{collect_results_collection}
\title{Collect the results from all recordings in a recording collection.}
\usage{
collect_results_collection(collection)
}
\arguments{
\item{collection}{A recording collection}
}
\value{
The analysis results as a data frame.
}
\description{
Get the analysis results, which are internally stored as a list of matrices,
from a recording and return them as a data frame.
}
\seealso{
Other recording: \code{\link{add_block}};
  \code{\link{add_segment_timestamp}};
  \code{\link{block_to_seconds}};
  \code{\link{collect_results}};
  \code{\link{create_block_simple}};
  \code{\link{create_block_structure}};
  \code{\link{cut_recording_s}};
  \code{\link{cut_recordings_s}};
  \code{\link{cut_recordings}};
  \code{\link{cut_recording}};
  \code{\link{find_recording_overlap}};
  \code{\link{load_recording}};
  \code{\link{new_recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording_get_zerotime}};
  \code{\link{recording_set_blocks}};
  \code{\link{recording_set_events}};
  \code{\link{recording_set_ibi}};
  \code{\link{recording_set_zerotime}};
  \code{\link{save_recording}}
}

