% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.block.simple}
\alias{create.block.simple}
\title{Create a simple block.}
\usage{
create.block.simple(starttime, starttype = "timestamp", stoptime,
  stoptype = "timestamp", tasktype = "none", part = NA, meas = NA,
  blockid = NA, subject = NA, casename = NA, recording = NULL)
}
\arguments{
\item{starttime}{The start time as a string.}

\item{starttype}{Either \code{time} or \code{timestamp}. Default is \code{timestamp}.}

\item{stoptime}{The stop time as a string.}

\item{stoptype}{Either \code{time} or \code{timestamp}. Default is \code{timestamp}.}

\item{tasktype}{The task type as a string. Default is \code{none}.}

\item{part}{An optional string describing the part.}

\item{meas}{An optional string describing the measurement}

\item{blockid}{An optional integer denoting the block ID.}

\item{subject}{An optional string describing the subject ID.}

\item{casename}{An optional string describing the casename.}

\item{recording}{An optional recording. If the recording is given,
the subject and casename are read from the recording and the
arguments for these are ignored.}
}
\value{
A block structure.
}
\description{
Create a simple block.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

