% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.block.structure}
\alias{create.block.structure}
\title{Create an empty block structure.}
\usage{
create.block.structure(recording, extra.factors = NULL)
}
\arguments{
\item{recording}{A recording.}

\item{extra.factors}{Extra factors to add to the block structure.}
}
\value{
An empty block structure (a data frame) with the following columns:
\describe{
\item{blockid}{A numeric ID for the block}
\item{casename}{The casename (string)}
\item{subject}{The subject ID (string)}
\item{part}{The part (string).}
\item{meas}{The measurement (string).}
\item{starttype}{The startttype. Can be either \code{time} (seconds from start of recording) or \code{timestamp} (an absolute timestamp in ISO-8601 format). }
\item{starttime}{The starttime; see starttype for details (string).}
\item{stoptype}{See starttype.}
\item{stoptime}{See starttime.}
\item{tasktype}{The task type (string).}
}
}
\description{
Create an empty block structure.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

