% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_block_simple}
\alias{create_block_simple}
\title{Create a simple block.}
\usage{
create_block_simple(starttime, starttype = "timestamp", stoptime,
  stoptype = "timestamp", tasktype = "none", part = NA, meas = NA,
  blockid = NA, subject = NA, casename = NA, recording = NULL)
}
\arguments{
\item{starttime}{The start time as a string.}

\item{starttype}{Either \code{time} or \code{timestamp}. Default is \code{timestamp}.}

\item{stoptime}{The stop time as a string.}

\item{stoptype}{Either \code{time} or \code{timestamp}. Default is \code{timestamp}.}

\item{tasktype}{The task type as a string. Default is \code{none}.}

\item{part}{An optional string describing the part.}

\item{meas}{An optional string describing the measurement}

\item{blockid}{An optional integer denoting the block ID.}

\item{subject}{An optional string describing the subject ID.}

\item{casename}{An optional string describing the casename.}

\item{recording}{An optional recording. If the recording is given,
the subject and casename are read from the recording and the
arguments for these are ignored.}
}
\value{
A block structure.
}
\description{
Create a simple block.
}
\seealso{
Other recording: \code{\link{add_block}};
  \code{\link{add_segment_timestamp}};
  \code{\link{block_to_seconds}};
  \code{\link{collect_results_collection}};
  \code{\link{collect_results}};
  \code{\link{create_block_structure}};
  \code{\link{cut_recording_s}};
  \code{\link{cut_recordings_s}};
  \code{\link{cut_recordings}};
  \code{\link{cut_recording}};
  \code{\link{find_recording_overlap}};
  \code{\link{load_recording}};
  \code{\link{new_recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording_get_zerotime}};
  \code{\link{recording_set_blocks}};
  \code{\link{recording_set_events}};
  \code{\link{recording_set_ibi}};
  \code{\link{recording_set_zerotime}};
  \code{\link{save_recording}}
}

