% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blocks.R
\name{create_block_structure}
\alias{create_block_structure}
\title{Create an empty block structure.}
\usage{
create_block_structure(recording, extra.factors = NULL)
}
\arguments{
\item{recording}{A recording.}

\item{extra.factors}{Extra factors to add to the block structure.}
}
\value{
An empty block structure (a data frame) with the following columns:
\describe{
\item{blockid}{A numeric ID for the block}
\item{casename}{The casename (string)}
\item{subject}{The subject ID (string)}
\item{part}{The part (string).}
\item{meas}{The measurement (string).}
\item{starttype}{The startttype. Can be either \code{time} (seconds from start of recording) or \code{timestamp} (an absolute timestamp in ISO-8601 format). }
\item{starttime}{The starttime; see starttype for details (string).}
\item{stoptype}{See starttype.}
\item{stoptime}{See starttime.}
\item{tasktype}{The task type (string).}
}
}
\description{
Create an empty block structure.
}
\seealso{
Other recording: \code{\link{add_block}},
  \code{\link{add_segment_timestamp}},
  \code{\link{block_to_seconds}},
  \code{\link{collect_results_collection}},
  \code{\link{collect_results}},
  \code{\link{create_block_simple}},
  \code{\link{cut_recording_s}},
  \code{\link{cut_recordings_s}},
  \code{\link{cut_recordings}},
  \code{\link{cut_recording}},
  \code{\link{find_recording_overlap}},
  \code{\link{load_recording}},
  \code{\link{new_recording}},
  \code{\link{read.block.event.data}},
  \code{\link{read.config.excel}},
  \code{\link{read.config.sqlite}},
  \code{\link{read.config}},
  \code{\link{read.dbtable.excel}},
  \code{\link{read.dbtable.sqlite}},
  \code{\link{recording_get_zerotime}},
  \code{\link{recording_set_blocks}},
  \code{\link{recording_set_events}},
  \code{\link{recording_set_ibi}},
  \code{\link{recording_set_zerotime}},
  \code{\link{save_recording}}
}

