% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cut.recording}
\alias{cut.recording}
\title{Cut a recording to the given time interval, given as timestamps.}
\usage{
\method{cut}{recording}(recording, ts = NULL)
}
\arguments{
\item{recording}{A recording.}

\item{ts}{A time interval given as a two-element list with two timestamps. The timestamps are given as strings in ISO-8601 format or as \code{POSIXct} timestamps.}
}
\value{
The recording, with all signals cut to the given time interval.
}
\description{
Also realign the time vector so that t = 0 is at the new start
time and recalculate the length of the recording.  All signals in
the recording are processed.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

