% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cut.recordings.s}
\alias{cut.recordings.s}
\title{Cut the recordings in a collection so that they represent the same time intervals.}
\usage{
\method{cut}{recordings.s}(collection, ts)
}
\arguments{
\item{collection}{A recording collection.}

\item{ts}{A time interval given as a two-element list where the elements correspond to the number of seconds from the zerotime of the recording.}
}
\value{
The recording collection, with all signals in all recordings cut to the given time interval.
}
\description{
Also realign the time vector so that t = 0 is at the new start
time. Also recalculate the length of the recordings.  All signals in
all recordings in the collection are processed.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

