% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cut_recordings_s}
\alias{cut_recordings_s}
\title{Cut the recordings in a collection so that they represent the same time intervals.}
\usage{
cut_recordings_s(collection, ts)
}
\arguments{
\item{collection}{A recording collection.}

\item{ts}{A time interval given as a two-element list where the elements correspond to the number of seconds from the zerotime of the recording.}
}
\value{
The recording collection, with all signals in all recordings cut to the given time interval.
}
\description{
Also realign the time vector so that t = 0 is at the new start
time. Also recalculate the length of the recordings.  All signals in
all recordings in the collection are processed.
}
\seealso{
Other recording: \code{\link{add_block}};
  \code{\link{add_segment_timestamp}};
  \code{\link{block_to_seconds}};
  \code{\link{collect_results_collection}};
  \code{\link{collect_results}};
  \code{\link{create_block_simple}};
  \code{\link{create_block_structure}};
  \code{\link{cut_recording_s}};
  \code{\link{cut_recordings}};
  \code{\link{cut_recording}};
  \code{\link{find_recording_overlap}};
  \code{\link{load_recording}};
  \code{\link{new_recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording_get_zerotime}};
  \code{\link{recording_set_blocks}};
  \code{\link{recording_set_events}};
  \code{\link{recording_set_ibi}};
  \code{\link{recording_set_zerotime}};
  \code{\link{save_recording}}
}

