% Generated by roxygen2 (4.0.2): do not edit by hand
\name{defactor}
\alias{defactor}
\title{Convert a factor to a string or numeric representation.}
\usage{
defactor(x, type = "numeric")
}
\arguments{
\item{x}{A factor}

\item{type}{A string indicating the type of the factor: "numeric" (default) or "string).}
}
\value{
A string representation of the factor.
}
\description{
Convert a factor to a string or numeric representation.
}
\seealso{
Other utilities: \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}

