% Generated by roxygen2 (4.0.2): do not edit by hand
\name{detect_artifacts_xu}
\alias{detect_artifacts_xu}
\title{Artifact detection from interbeat interval (IBI) time series.}
\usage{
detect_artifacts_xu(x)
}
\arguments{
\item{x}{A vector with the interbeat intervals (IBIs)}
}
\value{
A list with the values of the metrics.
}
\description{
Implements method presented by Xu and Schuckers.
Each RR interval is compared with the median of 25 surrounding RR
intervals and the last accepted RR interval. Only if both differences are
outside +- 20% the interval will be marked as an artifact.
}
\references{
Xu, X., Schuckers, S.: Automatic Detection of Artifacts in Heart
            Period Data. J Electrocardiol. 2001;34 Suppl:205-10.
}
\seealso{
Other HRV: \code{\link{plot_artifacts}};
  \code{\link{qrs_detect}}; \code{\link{qrs_interpolate}}
}

