% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ensure.path}
\alias{ensure.path}
\title{Ensure that a directory path exists. If the given path does not exist, it is created.}
\usage{
ensure.path(x, recursive = FALSE)
}
\arguments{
\item{x}{A path}

\item{recursive}{Boolean indicating whether the path should be created recursively. Default is FALSE.}
}
\value{
Nothing.
}
\description{
Ensure that a directory path exists. If the given path does not exist, it is created.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

