% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ensure_path}
\alias{ensure_path}
\title{Ensure that a directory path exists. If the given path does not exist, it is created.}
\usage{
ensure_path(x, recursive = FALSE)
}
\arguments{
\item{x}{A path}

\item{recursive}{Boolean indicating whether the path should be created recursively. Default is FALSE.}
}
\value{
Nothing.
}
\description{
Ensure that a directory path exists. If the given path does not exist, it is created.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}}
}

