% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract.segment.block}
\alias{extract.segment.block}
\title{Extract the segment of a signal corresponding to the given block ID.}
\usage{
extract.segment.block(recording, blockid, signal = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{blockid}{A numeric block ID.}

\item{signal}{A signal name found in the recording. Optional, default is NULL.}
}
\value{
A named list with the fields "data" and "t" from \code{signal} corresponding to the given block ID.
}
\description{
Extract the segment of a signal corresponding to the given block ID.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}

