% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract.segment.timestamp}
\alias{extract.segment.timestamp}
\title{Extract a segment of a signal corresponding to the given timestamps ts0 and ts1.}
\usage{
extract.segment.timestamp(recording, ts0, ts1 = NULL, signal = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{ts0}{Start timestamp or a two-element list giving the start and stop timestamps.}

\item{ts1}{Stop timestamp. Optional, default is NULL, in which case ind1 must a two-element list.}

\item{signal}{A signal name found in the recording. Optional, default is NULL.}
}
\value{
A named list with the fields "data" and "t" from \code{signal} corresponding to the given timestamps.
}
\description{
Extract a segment of a signal corresponding to the given timestamps ts0 and ts1.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

