% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_segment}
\alias{extract_segment}
\title{Extract a segment of a signal corresponding to the given indices.}
\usage{
extract_segment(recording, ind1, ind2 = NULL, signal = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{ind1}{Start index or a two-element list giving the start and stop index.}

\item{ind2}{Stop index. Optional, default is NULL, in which case ind1 must a two-element list.}

\item{signal}{A signal name found in the recording. Optional, default is NULL.}
}
\value{
A named list with the fields "data" and "t" from \code{signal} corresponding to the given indices.
}
\description{
Extract a segment of a signal corresponding to the given indices.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{hr_from_ibi}}; \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}};
  \code{\link{timestamp_to_str}}
}

