% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_segment_s}
\alias{extract_segment_s}
\title{Extract a segment of a signal corresponding to the given times t0 and t1, given as seconds from the start of the recording.}
\usage{
extract_segment_s(recording, t0, t1, signal = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{t0}{Start time (in seconds) or a two-element list giving the start and stop time.}

\item{t1}{Stop time. Optional, default is NULL, in which case ind1 must a two-element list.}

\item{signal}{A signal name found in the recording. Optional, default is NULL.}
}
\value{
A named list with the fields "data" and "t" from \code{signal} corresponding to the given times.
}
\description{
Extract a segment of a signal corresponding to the given times t0 and t1, given as seconds from the start of the recording.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}}
}

