% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_segment_timestamp}
\alias{extract_segment_timestamp}
\title{Extract a segment of a signal corresponding to the given timestamps ts0 and ts1.}
\usage{
extract_segment_timestamp(recording, ts0, ts1 = NULL, signal = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{ts0}{Start timestamp or a two-element list giving the start and stop timestamps.}

\item{ts1}{Stop timestamp. Optional, default is NULL, in which case ind1 must a two-element list.}

\item{signal}{A signal name found in the recording. Optional, default is NULL.}
}
\value{
A named list with the fields "data" and "t" from \code{signal} corresponding to the given timestamps.
}
\description{
Extract a segment of a signal corresponding to the given timestamps ts0 and ts1.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{hr_from_ibi}}; \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}};
  \code{\link{timestamp_to_str}}
}

