% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\name{find_recording_overlap}
\alias{find_recording_overlap}
\title{Given a recording collection, find the overlapping times of all the recordings.}
\usage{
find_recording_overlap(collection)
}
\arguments{
\item{collection}{A recording collection (a list of recordings).}
}
\value{
A list containing the start and stop times fo
}
\description{
Given a recording collection, find the overlapping times of all the recordings.
}
\seealso{
Other recording: \code{\link{add_block}},
  \code{\link{add_segment_timestamp}},
  \code{\link{block_to_seconds}},
  \code{\link{collect_results_collection}},
  \code{\link{collect_results}},
  \code{\link{create_block_simple}},
  \code{\link{create_block_structure}},
  \code{\link{cut_recording_s}},
  \code{\link{cut_recordings_s}},
  \code{\link{cut_recordings}},
  \code{\link{cut_recording}},
  \code{\link{load_recording}},
  \code{\link{new_recording}},
  \code{\link{read.block.event.data}},
  \code{\link{read.config.excel}},
  \code{\link{read.config.sqlite}},
  \code{\link{read.config}},
  \code{\link{read.dbtable.excel}},
  \code{\link{read.dbtable.sqlite}},
  \code{\link{recording_get_zerotime}},
  \code{\link{recording_set_blocks}},
  \code{\link{recording_set_events}},
  \code{\link{recording_set_ibi}},
  \code{\link{recording_set_zerotime}},
  \code{\link{save_recording}}
}

