% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.result.row}
\alias{generate.result.row}
\title{Generate a result row to be used as a template for the results, based on the given block.}
\usage{
generate.result.row(block)
}
\arguments{
\item{block}{A block structure.}
}
\value{
A data frame row to be used as a template for results.
}
\description{
Generate a result row to be used as a template for the results, based on the given block.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

