% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.segments.from.block}
\alias{generate.segments.from.block}
\title{Generate analysis segments given a block.}
\usage{
generate.segments.from.block(block, settings, tolerance = 1)
}
\arguments{
\item{block}{A block structure}

\item{settings}{The settings structure used for HRV analysis.}

\item{tolerance}{The tolerance of the last segment length in seconds. Optional, default is 1.}
}
\value{
A two-column matrix with the start time (in seconds) of
each segment in the first column and the stop time (in seconds) in
the second column.
}
\description{
If \code{segmentlength} is zero, the entire block length is used
and no segments are created. The tolerance is given in seconds, and
defaults to 1 second. The tolerance affects the last segment. If
the last segment would, e.g., extend past the data, the tolerance
parameter can be used to adjust when a slightly-shorter segment is
accepted, and when the entire segment must be discarded. The
segment length and overlap are read from the \code{settings}
structure.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

