% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.segments}
\alias{generate.segments}
\title{Generate analysis segments given a start and stop time.}
\usage{
generate.segments(time.start, time.stop, segment.length, segment.overlap,
  tolerance = 1)
}
\arguments{
\item{time.start}{A start time in seconds.}

\item{time.stop}{A stop time in seconds.}

\item{segment.length}{The segment length}

\item{segment.overlap}{The segment overlap}

\item{tolerance}{The tolerance of the last segment length in seconds. Optional, default is 1.}

\item{blockid}{A numeric block ID.}
}
\value{
A two-column matrix with the start time (in seconds) of
each segment in the first column and the stop time (in seconds) in
the second column.
}
\description{
If \code{segmentlength} is zero, the entire block length is used
and no segments are created. The tolerance is given in seconds, and
defaults to 1 second. The tolerance affects the last segment. If
the last segment would, e.g., extend past the data, the tolerance
parameter can be used to adjust when a slightly-shorter segment is
accepted, and when the entire segment must be discarded.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

