% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_result_row}
\alias{generate_result_row}
\title{Generate a result row to be used as a template for the results, based on the given block.}
\usage{
generate_result_row(block)
}
\arguments{
\item{block}{A block structure.}
}
\value{
A data frame row to be used as a template for results.
}
\description{
Generate a result row to be used as a template for the results, based on the given block.
}
\seealso{
Other utilities: \code{\link{defactor}},
  \code{\link{ensure_path}},
  \code{\link{extract_segment_block}},
  \code{\link{extract_segment_s}},
  \code{\link{extract_segment_timestamp}},
  \code{\link{extract_segment}},
  \code{\link{generate_segments_from_block}},
  \code{\link{generate_segments}},
  \code{\link{get_file_extension}},
  \code{\link{get_filename_parts}},
  \code{\link{hr_from_ibi}}, \code{\link{ibi_from_ecg}},
  \code{\link{integrate_power}},
  \code{\link{num_to_timestamp}},
  \code{\link{replicate_df_row}},
  \code{\link{seconds_to_indices}},
  \code{\link{str_to_timestamp}},
  \code{\link{timeintervals_to_seconds}},
  \code{\link{timestamp_to_seconds}},
  \code{\link{timestamp_to_str}}
}

