% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_segments_from_block}
\alias{generate_segments_from_block}
\title{Generate analysis segments given a block.}
\usage{
generate_segments_from_block(block, settings, tolerance = 1)
}
\arguments{
\item{block}{A block structure}

\item{settings}{The settings structure used for HRV analysis.}

\item{tolerance}{The tolerance of the last segment length in seconds. Optional, default is 1.}
}
\value{
A two-column matrix with the start time (in seconds) of
each segment in the first column and the stop time (in seconds) in
the second column.
}
\description{
If \code{segmentlength} is zero, the entire block length is used
and no segments are created. The tolerance is given in seconds, and
defaults to 1 second. The tolerance affects the last segment. If
the last segment would, e.g., extend past the data, the tolerance
parameter can be used to adjust when a slightly-shorter segment is
accepted, and when the entire segment must be discarded. The
segment length and overlap are read from the \code{settings}
structure.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{hr_from_ibi}}; \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}}
}

