% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.filename.parts}
\alias{get.filename.parts}
\title{Separate the filename and the file extenions}
\usage{
get.filename.parts(filename)
}
\arguments{
\item{filename}{A filename with an extension.}
}
\value{
A list containing the filename and the file extension.
}
\description{
Separate the filename and the file extenions
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

