% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_filename_parts}
\alias{get_filename_parts}
\title{Separate the filename and the file extenions}
\usage{
get_filename_parts(filename)
}
\arguments{
\item{filename}{A filename with an extension.}
}
\value{
A list containing the filename and the file extension.
}
\description{
Separate the filename and the file extenions
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{hr_from_ibi}}; \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{num_to_timestamp}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}};
  \code{\link{timestamp_to_str}}
}
\keyword{internal}

