% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hr_from_ibi}
\alias{hr_from_ibi}
\title{Add a time series with the instantaneous heart rate to a recording.}
\usage{
hr_from_ibi(recording, signal = "ibi")
}
\arguments{
\item{recording}{A recording structure.}

\item{signal}{A signal name containing interbeat intervals in the recording. Optional, default is "ibi".}
}
\value{
The recording structure with the new signal "hr" added.
}
\description{
Add a time series with the instantaneous heart rate to a recording.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}}
}

