% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.band.custom}
\alias{ibi.band.custom}
\title{Calculate the power in custom HRV frequency bands.}
\usage{
ibi.band.custom(spec, settings)
}
\arguments{
\item{spec}{The Lomb-Scargle spectrum estimated from the IBI series using the function lombscargle.}

\item{settings}{The settings structure used for HRV analysis with the following fields:
\describe{
\item{$frequencydomain$parameters$band.lower}{A list with the lower frequency bound for the custom bands.}
\item{$frequencydomain$parameters$band.upper}{A list with the upper frequency bound for the custom bands.}
\item{$frequencydomain$parameters$band.names.custom}{A list with the names of the custom bands.}
}}
}
\value{
The frequencies in the custom bands.
}
\description{
Calculate the power in custom HRV frequency bands.
}
\seealso{
Other HRV frequency-domain: \code{\link{analyse.frequencydomain.helper}};
  \code{\link{analyse.frequencydomain}};
  \code{\link{ibi.band.standard}};
  \code{\link{lombscargle}}
}

