% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.from.ecg}
\alias{ibi.from.ecg}
\title{Given an ECG signal, form the IBI series}
\usage{
ibi.from.ecg(recording, signal = "ECG", interpolate.qrs = TRUE)
}
\arguments{
\item{recording}{A recording structure.}

\item{signal}{A signal name found in the recording. Optional, default is "ECG".}

\item{interpolate.qrs}{Should the position of the R-peak be interpolated. See \code{\link{qrs.interpolate}} for details.}
}
\value{
The recording structure with the new signals "ibi" and "ibi.amp" added.
}
\description{
Given an ECG signal, form the IBI series
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}

