% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.mean}
\alias{ibi.mean}
\title{Calculate the average of interbeat intervals or heart rate.
The time series is in miliseconds}
\usage{
ibi.mean(ibi, type = "ibi")
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{type}{String denoting whether to calculate the average interbeat interval (type = "ibi", default) or the average heart rate (type = "hr").}
}
\value{
The average interbeat interval or the average heart rate.
}
\description{
Calculate the average of interbeat intervals or heart rate.
The time series is in miliseconds
}
\seealso{
Other HRV time-domain: \code{\link{analyse.timedomain}};
  \code{\link{ibi.nnx}}; \code{\link{ibi.pnnx.helper}};
  \code{\link{ibi.pnnx}}; \code{\link{ibi.rmssd}};
  \code{\link{ibi.stdev}}; \code{\link{ibi.var}}
}
\keyword{internal}

