% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.pnnx.helper}
\alias{ibi.pnnx.helper}
\title{Helper function for calculating the pNNx value.}
\usage{
ibi.pnnx.helper(ibi, pnnx.list, type = "percentage")
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{pnnx.list}{A list of numbers with the limits in milliseconds (the x in pNNx and NNx) for calculating the pNNx and NNx.}

\item{type}{A string denoting what to calculate; type =
"percentage" to calculate the percentage of NN values (default), or
type = "number" to calculate the number of NN intervals.}
}
\value{
A list with the pNNx or NNx values.
}
\description{
Helper function for calculating the pNNx value.
}
\seealso{
Other HRV time-domain: \code{\link{analyse.timedomain}};
  \code{\link{ibi.mean}}; \code{\link{ibi.nnx}};
  \code{\link{ibi.pnnx}}; \code{\link{ibi.rmssd}};
  \code{\link{ibi.stdev}}; \code{\link{ibi.var}}
}
\keyword{internal}

