% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.pnnx}
\alias{ibi.pnnx}
\title{Calculate the NNx value.}
\usage{
ibi.pnnx(ibi, x = 50)
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{x}{A number with the limit in milliseconds (the x in pNNx) for calculating the NNx; the percentage of interbeat intervals differing more than x ms. Default is 50.}
}
\value{
The pNNx value.
}
\description{
Calculate the NNx value.
}
\seealso{
Other HRV time-domain: \code{\link{analyse.timedomain}};
  \code{\link{ibi.mean}}; \code{\link{ibi.nnx}};
  \code{\link{ibi.pnnx.helper}}; \code{\link{ibi.rmssd}};
  \code{\link{ibi.stdev}}; \code{\link{ibi.var}}
}
\keyword{internal}

