% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.sd2}
\alias{ibi.sd2}
\title{Determine the SD2 metric from a Poincare plot.}
\usage{
ibi.sd2(ibi)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs).}
}
\value{
The value of the SD2 metric.
}
\description{
Determine the SD2 metric from a Poincare plot.
}
\seealso{
Other HRV geometric: \code{\link{analyse.geometric.helper}};
  \code{\link{analyse.geometric}}; \code{\link{ibi.sd1}};
  \code{\link{ibi.tinn}};
  \code{\link{ibi.triangular.index}}
}

