% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.stdev}
\alias{ibi.stdev}
\title{Calculate the standard deviation of interbeat intervals or heart rate.}
\usage{
ibi.stdev(ibi, type = "ibi")
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{type}{String denoting whether to calculate the average interbeat interval (type = "ibi", default) or the average heart rate (type = "hr").}
}
\value{
The standard deviation of interbeat intervals or heart rate.
}
\description{
Calculate the standard deviation of interbeat intervals or heart rate.
}
\seealso{
Other HRV time-domain: \code{\link{analyse.timedomain}};
  \code{\link{ibi.mean}}; \code{\link{ibi.nnx}};
  \code{\link{ibi.pnnx.helper}}; \code{\link{ibi.pnnx}};
  \code{\link{ibi.rmssd}}; \code{\link{ibi.var}}
}
\keyword{internal}

