% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.tinn}
\alias{ibi.tinn}
\title{Determine the HRV Triangular Interpolation Index (TINN)
as defined in HRV Guidelines.}
\usage{
ibi.tinn(ibi, binwidth = 128)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{binwidth}{The reciprocal of the width of the histogram bins. Default is 128.}
}
\value{
The HRV triangular interpolat index (TINN).
}
\description{
Determine the HRV Triangular Interpolation Index (TINN)
as defined in HRV Guidelines.
}
\seealso{
Other HRV geometric: \code{\link{analyse.geometric.helper}};
  \code{\link{analyse.geometric}}; \code{\link{ibi.sd1}};
  \code{\link{ibi.sd2}}; \code{\link{ibi.triangular.index}}
}

