% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi.var}
\alias{ibi.var}
\title{Calculate the variance of interbeat intervals.}
\usage{
ibi.var(ibi)
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}
}
\value{
The variance of interbeat intervals.
}
\description{
Calculate the variance of interbeat intervals.
}
\seealso{
Other HRV time-domain: \code{\link{analyse.timedomain}};
  \code{\link{ibi.mean}}; \code{\link{ibi.nnx}};
  \code{\link{ibi.pnnx.helper}}; \code{\link{ibi.pnnx}};
  \code{\link{ibi.rmssd}}; \code{\link{ibi.stdev}}
}
\keyword{internal}

