% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis_hrv_frequencydomain.R
\name{ibi_band_custom}
\alias{ibi_band_custom}
\title{Calculate the power in custom HRV frequency bands.}
\usage{
ibi_band_custom(spec, settings)
}
\arguments{
\item{spec}{The Lomb-Scargle spectrum estimated from the IBI series using the function lombscargle.}

\item{settings}{The settings structure used for HRV analysis with the following fields:
\describe{
\item{$frequencydomain$parameters$band.lower}{A list with the lower frequency bound for the custom bands.}
\item{$frequencydomain$parameters$band.upper}{A list with the upper frequency bound for the custom bands.}
\item{$frequencydomain$parameters$band.names.custom}{A list with the names of the custom bands.}
}}
}
\value{
The frequencies in the custom bands.
}
\description{
Calculate the power in custom HRV frequency bands.
}
\seealso{
Other HRV frequency-domain: \code{\link{analyse_frequencydomain_helper}};
  \code{\link{analyse_frequencydomain}};
  \code{\link{ibi_band_standard}};
  \code{\link{lombscargle}}
}

