% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hrv_frequencydomain.R
\name{ibi_band_standard}
\alias{ibi_band_standard}
\title{Calculate the power in the standard HRV frequency bands (VLF, LF, HF).}
\usage{
ibi_band_standard(spec, settings)
}
\arguments{
\item{spec}{The Lomb-Scargle spectrum estimated from the IBI series using the function lombscargle.}

\item{settings}{The settings structure used for HRV analysis with the following fields:
\describe{
\item{$frequencydomain$parameters$f.limits}{A two-element list with the upper and lower frequency bounds for the spectrum estimation.}
\item{$frequencydomain$parameters$band.vlf)}{A two-element list with the upper and lower frequency bounds for the VLF band.}
\item{$frequencydomain$parameters$band.lf)}{A two-element list with the upper and lower frequency bounds for the LF band.}
\item{$frequencydomain$parameters$band.hf)}{A two-element list with the upper and lower frequency bounds for the HF band.}
}}
}
\value{
The input structure res with the following fields added:
}
\description{
Calculate the power in the standard HRV frequency bands (VLF, LF, HF).
}
\seealso{
Other HRV frequency-domain: \code{\link{analyse_frequencydomain_helper}},
  \code{\link{analyse_frequencydomain}},
  \code{\link{ibi_band_custom}}, \code{\link{lombscargle}}
}

