% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi_entropy}
\alias{ibi_entropy}
\title{Calculate entropy of the IBI series}
\usage{
ibi_entropy(x, embedding_dimension = 2, type = NULL)
}
\arguments{
\item{x}{A vector with the interbeat intervals (IBIs).}

\item{type}{The type of entropy as a string: \code{approximate} or \code{sample}}

\item{embedding.dimension}{The embedding dimension of the entropy. The default is 2.}
}
\value{
The entropy.
}
\description{
Calculate entropy of the IBI series
}
\seealso{
Other HRV nonlinear: \code{\link{analyse_nonlinear_helper}};
  \code{\link{analyse_nonlinear}}
}

