% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ibi_from_ecg}
\alias{ibi_from_ecg}
\title{Given an ECG signal, form the IBI series}
\usage{
ibi_from_ecg(recording, signal = "ECG", interpolate.qrs = TRUE, ...)
}
\arguments{
\item{recording}{A recording structure.}

\item{signal}{A signal name found in the recording. Optional, default is "ECG".}

\item{interpolate.qrs}{Should the position of the R-peak be interpolated. See \code{\link{qrs_interpolate}} for details.}

\item{...}{Additional arguments to \code{\link{qrs_detect}}.}
}
\value{
The recording structure with the new signals "ibi" and "ibi.amp" added.
}
\description{
Given an ECG signal, form the IBI series
}
\seealso{
Other utilities: \code{\link{defactor}},
  \code{\link{ensure_path}},
  \code{\link{extract_segment_block}},
  \code{\link{extract_segment_s}},
  \code{\link{extract_segment_timestamp}},
  \code{\link{extract_segment}},
  \code{\link{generate_result_row}},
  \code{\link{generate_segments_from_block}},
  \code{\link{generate_segments}},
  \code{\link{get_file_extension}},
  \code{\link{get_filename_parts}},
  \code{\link{hr_from_ibi}}, \code{\link{integrate_power}},
  \code{\link{num_to_timestamp}},
  \code{\link{replicate_df_row}},
  \code{\link{seconds_to_indices}},
  \code{\link{str_to_timestamp}},
  \code{\link{timeintervals_to_seconds}},
  \code{\link{timestamp_to_seconds}},
  \code{\link{timestamp_to_str}}
}

