% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hrv_timedomain.R
\name{ibi_mean}
\alias{ibi_mean}
\title{Calculate the average of interbeat intervals or heart rate.
The time series is in miliseconds}
\usage{
ibi_mean(ibi, type = "ibi")
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{type}{String denoting whether to calculate the average interbeat interval (type = "ibi", default) or the average heart rate (type = "hr").}
}
\value{
The average interbeat interval or the average heart rate.
}
\description{
Calculate the average of interbeat intervals or heart rate.
The time series is in miliseconds
}
\seealso{
Other HRV time-domain: \code{\link{analyse_timedomain}},
  \code{\link{ibi_max}}, \code{\link{ibi_median}},
  \code{\link{ibi_min}}, \code{\link{ibi_nnx}},
  \code{\link{ibi_pnnx_helper}}, \code{\link{ibi_pnnx}},
  \code{\link{ibi_rmssd}}, \code{\link{ibi_stdev}},
  \code{\link{ibi_var}}
}
\keyword{internal}

