% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hrv_timedomain.R
\name{ibi_median}
\alias{ibi_median}
\title{Calculate the median of interbeat intervals or heart rate.
The time series is in miliseconds}
\usage{
ibi_median(ibi, type = "ibi")
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{type}{String denoting whether to calculate the median interbeat interval (type = "ibi", default) or the median heart rate (type = "hr").}
}
\value{
The median interbeat interval or the median heart rate.
}
\description{
Calculate the median of interbeat intervals or heart rate.
The time series is in miliseconds
}
\seealso{
Other HRV time-domain: \code{\link{analyse_timedomain}},
  \code{\link{ibi_max}}, \code{\link{ibi_mean}},
  \code{\link{ibi_min}}, \code{\link{ibi_nnx}},
  \code{\link{ibi_pnnx_helper}}, \code{\link{ibi_pnnx}},
  \code{\link{ibi_rmssd}}, \code{\link{ibi_stdev}},
  \code{\link{ibi_var}}
}
\keyword{internal}

