% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi_nnx}
\alias{ibi_nnx}
\title{Calculate the NNx value.}
\usage{
ibi_nnx(ibi, x = 50)
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{x}{A number with the limit in milliseconds (the x in NNx) for calculating the NNx; the number of interbeat intervals differing more than x ms. Default is 50.}
}
\value{
The NNx value.
}
\description{
Calculate the NNx value.
}
\seealso{
Other HRV time-domain: \code{\link{analyse_timedomain}};
  \code{\link{ibi_mean}}; \code{\link{ibi_pnnx_helper}};
  \code{\link{ibi_pnnx}}; \code{\link{ibi_rmssd}};
  \code{\link{ibi_stdev}}; \code{\link{ibi_var}}
}
\keyword{internal}

