% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hrv_timedomain.R
\name{ibi_pnnx_helper}
\alias{ibi_pnnx_helper}
\title{Helper function for calculating the pNNx value.}
\usage{
ibi_pnnx_helper(ibi, pnnx.list, type = "percentage")
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{pnnx.list}{A list of numbers with the limits in milliseconds (the x in pNNx and NNx) for calculating the pNNx and NNx.}

\item{type}{A string denoting what to calculate; type =
"percentage" to calculate the percentage of NN values (default), or
type = "number" to calculate the number of NN intervals.}
}
\value{
A list with the pNNx or NNx values.
}
\description{
Helper function for calculating the pNNx value.
}
\seealso{
Other HRV time-domain: \code{\link{analyse_timedomain}},
  \code{\link{ibi_max}}, \code{\link{ibi_mean}},
  \code{\link{ibi_median}}, \code{\link{ibi_min}},
  \code{\link{ibi_nnx}}, \code{\link{ibi_pnnx}},
  \code{\link{ibi_rmssd}}, \code{\link{ibi_stdev}},
  \code{\link{ibi_var}}
}
\keyword{internal}

