% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi_rmssd}
\alias{ibi_rmssd}
\title{Calculate the root-mean squared of successive differences (RMSSD).}
\usage{
ibi_rmssd(ibi)
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}
}
\value{
The RMSSD value.
}
\description{
Calculate the root-mean squared of successive differences (RMSSD).
}
\seealso{
Other HRV time-domain: \code{\link{analyse_timedomain}};
  \code{\link{ibi_max}}; \code{\link{ibi_mean}};
  \code{\link{ibi_median}}; \code{\link{ibi_min}};
  \code{\link{ibi_nnx}}; \code{\link{ibi_pnnx_helper}};
  \code{\link{ibi_pnnx}}; \code{\link{ibi_stdev}};
  \code{\link{ibi_var}}
}
\keyword{internal}

