% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hrv_geometric.R
\name{ibi_sd1}
\alias{ibi_sd1}
\title{Determine the SD1 metric from a Poincare plot.}
\usage{
ibi_sd1(ibi)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs).}
}
\value{
The value of the SD1 metric.
}
\description{
Determine the SD1 metric from a Poincare plot.
}
\seealso{
Other HRV geometric: \code{\link{analyse_geometric_helper}},
  \code{\link{analyse_geometric}}, \code{\link{ibi_sd2}},
  \code{\link{ibi_tinn}},
  \code{\link{ibi_triangular_index}}
}

