% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi_sd2}
\alias{ibi_sd2}
\title{Determine the SD2 metric from a Poincare plot.}
\usage{
ibi_sd2(ibi)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs).}
}
\value{
The value of the SD2 metric.
}
\description{
Determine the SD2 metric from a Poincare plot.
}
\seealso{
Other HRV geometric: \code{\link{analyse_geometric_helper}};
  \code{\link{analyse_geometric}}; \code{\link{ibi_sd1}};
  \code{\link{ibi_tinn}};
  \code{\link{ibi_triangular_index}}
}

