% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi_stdev}
\alias{ibi_stdev}
\title{Calculate the standard deviation of interbeat intervals or heart rate.}
\usage{
ibi_stdev(ibi, type = "ibi")
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}

\item{type}{String denoting whether to calculate the average interbeat interval (type = "ibi", default) or the average heart rate (type = "hr").}
}
\value{
The standard deviation of interbeat intervals or heart rate.
}
\description{
Calculate the standard deviation of interbeat intervals or heart rate.
}
\seealso{
Other HRV time-domain: \code{\link{analyse_timedomain}};
  \code{\link{ibi_max}}; \code{\link{ibi_mean}};
  \code{\link{ibi_median}}; \code{\link{ibi_min}};
  \code{\link{ibi_nnx}}; \code{\link{ibi_pnnx_helper}};
  \code{\link{ibi_pnnx}}; \code{\link{ibi_rmssd}};
  \code{\link{ibi_var}}
}
\keyword{internal}

