% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi_tinn}
\alias{ibi_tinn}
\title{Determine the HRV Triangular Interpolation Index (TINN)
as defined in HRV Guidelines.}
\usage{
ibi_tinn(ibi, binwidth = 128)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{binwidth}{The reciprocal of the width of the histogram bins. Default is 128.}
}
\value{
The HRV triangular interpolat index (TINN).
}
\description{
Determine the HRV Triangular Interpolation Index (TINN)
as defined in HRV Guidelines.
}
\seealso{
Other HRV geometric: \code{\link{analyse_geometric_helper}};
  \code{\link{analyse_geometric}}; \code{\link{ibi_sd1}};
  \code{\link{ibi_sd2}}; \code{\link{ibi_triangular_index}}
}

