% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ibi_triangular_index}
\alias{ibi_triangular_index}
\title{Determine the HRV Triangular index as defined in HRV Guidelines:}
\usage{
ibi_triangular_index(ibi, binwidth = 128)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs).}

\item{binwidth}{The reciprocal of the width of the histogram bins. Default is 128.}
}
\value{
The HRV triangular index.
}
\description{
(1) IBI histogram using bins of width 1 / 128 seconds and.
(2) Determine the number of intervals in the modal bin
(3) Triangular index:
(total number of NN intervals) /(number of NN intervals in the
modal bin).
}
\seealso{
Other HRV geometric: \code{\link{analyse_geometric_helper}};
  \code{\link{analyse_geometric}}; \code{\link{ibi_sd1}};
  \code{\link{ibi_sd2}}; \code{\link{ibi_tinn}}
}

