% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis_hrv_timedomain.R
\name{ibi_var}
\alias{ibi_var}
\title{Calculate the variance of interbeat intervals.}
\usage{
ibi_var(ibi)
}
\arguments{
\item{ibi}{An array with interbeat intervals in milliseconds.}
}
\value{
The variance of interbeat intervals.
}
\description{
Calculate the variance of interbeat intervals.
}
\seealso{
Other HRV time-domain: \code{\link{analyse_timedomain}};
  \code{\link{ibi_max}}; \code{\link{ibi_mean}};
  \code{\link{ibi_median}}; \code{\link{ibi_min}};
  \code{\link{ibi_nnx}}; \code{\link{ibi_pnnx_helper}};
  \code{\link{ibi_pnnx}}; \code{\link{ibi_rmssd}};
  \code{\link{ibi_stdev}}
}
\keyword{internal}

