% Generated by roxygen2 (4.0.2): do not edit by hand
\name{integrate.power}
\alias{integrate.power}
\title{Integrate using the trapezoid rule}
\usage{
integrate.power(f, p, fmin, fmax)
}
\arguments{
\item{f}{A vector of frequencies.}

\item{p}{A vector of power values corresponding to the frequency values.}

\item{fmin}{The minimum frequency value.}

\item{fmax}{The maximum frequency value.}
}
\value{
The integrated power.
}
\description{
Integrate using the trapezoid rule
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}

