% Generated by roxygen2 (4.0.2): do not edit by hand
\name{integrate_power}
\alias{integrate_power}
\title{Integrate using the trapezoid rule}
\usage{
integrate_power(f, p, fmin, fmax)
}
\arguments{
\item{f}{A vector of frequencies.}

\item{p}{A vector of power values corresponding to the frequency values.}

\item{fmin}{The minimum frequency value.}

\item{fmax}{The maximum frequency value.}
}
\value{
The integrated power.
}
\description{
Integrate using the trapezoid rule
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{ibi_from_ecg}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}}
}

