% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moving.average}
\alias{moving.average}
\title{Calculate the moving average of a vector.}
\usage{
moving.average(x, n = 10)
}
\arguments{
\item{x}{A vector with data.}

\item{n}{Window width in samples.}
}
\value{
The data in x, averaged in a window of width n.
}
\description{
Calculate the moving average of a vector.
}

