% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrs_detect.R
\name{moving_average}
\alias{moving_average}
\title{Calculate the moving average of a vector.}
\usage{
moving_average(x, n = 10)
}
\arguments{
\item{x}{A vector with data.}

\item{n}{Window width in samples.}
}
\value{
The data in x, averaged in a window of width n.
}
\description{
Calculate the moving average of a vector.
}

