% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{num_to_timestamp}
\alias{num_to_timestamp}
\title{Convert a numeric value to a timestamp
Assumes the numeric values come from a call to as.numeric() on a POSIXct with
time zone tz="UTC"!}
\usage{
num_to_timestamp(tn)
}
\arguments{
\item{tn}{A numeric vector representing time in seconds since the epoch ‘1970-01-01 00:00.00 UTC’}
}
\value{
A vector of timestamps of class POSIXct
}
\description{
Convert a numeric value to a timestamp
Assumes the numeric values come from a call to as.numeric() on a POSIXct with
time zone tz="UTC"!
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{hr_from_ibi}}; \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}};
  \code{\link{timestamp_to_str}}
}

