% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.artifacts}
\alias{plot.artifacts}
\title{Visualize artifacts in IBI time series.}
\usage{
\method{plot}{artifacts}(ibi, ibi.t)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs)}

\item{ibi.t}{A vector with the times of occurrence of the interbeat intervals (IBIs)}
}
\value{
Nothing.
}
\description{
Artifact detection is performed using the method presented by Xu and Schuckers
\code{\link{detect.artifacts.xu}} and the result is visualized.
}
\seealso{
Other HRV: \code{\link{detect.artifacts.xu}};
  \code{\link{qrs.detect}}; \code{\link{qrs.interpolate}}
}

