% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.poincare}
\alias{plot.poincare}
\title{Produce a Poincare plot.}
\usage{
\method{plot}{poincare}(ibi)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs).}
}
\value{
Nothing
}
\description{
Produce a Poincare plot.
}

