% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_ecg_r_peak}
\alias{plot_ecg_r_peak}
\title{Produce a multipage PDF file with plots showing the entire raw ECG signal together with markers for the R-peaks.}
\usage{
plot_ecg_r_peak(recording, filename, n = 60, signal = "ECG",
  signal.color = "blue", signal2 = NULL, signal2.color = "red",
  signal2.pch = 10, signal2.cex = 2)
}
\arguments{
\item{recording}{A recording structure.}

\item{filename}{The full path and filename in which to save the figure. The format is always PDF.}

\item{n}{The number of seconds of ECG data to show per row in the plot (default is 60 seconds).}

\item{signal}{The name of the ECG signal in the recording structure (default is 'ECG').}

\item{signal.color}{The line color of the ECG signal. Default is 'blue'.}

\item{signal2}{The name of the second signal in the recording structure (default is NULL, but 'ibi.amp' would be a typical choice).}

\item{signal2.color}{The color of the points for the secondary signal. Default is 'red'.}

\item{signal2.pch}{The point type used to plot the secondary signal. Default is 10.}

\item{signal2.cex}{The point size used to plot the secondary signal. Default is 2.}
}
\value{
Nothing. Saves the figure as multipage PDF.
}
\description{
Plot the raw ECG data together with R-peaks. This allows the quality of the R-peak detection to be checked visually for the entire recording.
}
\seealso{
Other visualizations: \code{\link{plot_all_blocks}};
  \code{\link{plot_block}};
  \code{\link{plot_check_rr_detection}};
  \code{\link{plot_metric}}; \code{\link{plot_spectrum}};
  \code{\link{shade_frequency_band}}
}

