% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_metric}
\alias{plot_metric}
\title{Plot a metric.}
\usage{
plot_metric(recording, metric, filename, new.plot = TRUE, blockid = NULL)
}
\arguments{
\item{recording}{A recording structure. That must contain the results structure.}

\item{metric}{The metric to plot. The metric must be found in the results structure.}

\item{filename}{The full path and filename in which to save the figure. If NULl (default), the figure is not saved. The format of the figure (PDF or PNG) is automatically determined from the filename extension. The size of the figure is an A4 in landscape mode.}

\item{new.plot}{Boolean indicating whether a new plot window should be opened (optional, default is TRUE).}

\item{blockid}{The ID of the measurement block (optional, default is NULL, which corresponds to plotting all data).}
}
\value{
Nothing. Produces a plot or saves the figure.
}
\description{
Plot a metric.
}
\seealso{
Other visualizations: \code{\link{plot_all_blocks}},
  \code{\link{plot_block}},
  \code{\link{plot_check_rr_detection}},
  \code{\link{plot_ecg_r_peak}},
  \code{\link{plot_spectrum}},
  \code{\link{shade_frequency_band}}
}

