% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hrv_geometric.R
\name{plot_poincare}
\alias{plot_poincare}
\title{Produce a Poincare plot.}
\usage{
plot_poincare(ibi)
}
\arguments{
\item{ibi}{A vector with the interbeat intervals (IBIs).}
}
\value{
Nothing
}
\description{
Produce a Poincare plot.
}

