% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot the HRV frequency spectrum.}
\usage{
plot_spectrum(recording, signal = "ibi")
}
\arguments{
\item{recording}{A recording structure. This structure must also contain the settings structure.}
}
\value{
Nothing. Produces a plot.
}
\description{
Plot the HRV frequency spectrum.
}
\seealso{
Other visualizations: \code{\link{plot_all_blocks}};
  \code{\link{plot_block}};
  \code{\link{plot_check_rr_detection}};
  \code{\link{plot_ecg_r_peak}}; \code{\link{plot_metric}};
  \code{\link{shade_frequency_band}}
}

