% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qrs.detect}
\alias{qrs.detect}
\title{R-peak detection}
\usage{
qrs.detect(ecg, samplingrate, interpolate = TRUE)
}
\arguments{
\item{ecg}{A vector with the ECG signal}

\item{samplingrate}{The sampling rate of the ECG signal}

\item{interpolate}{Should the position of the R-peak be interpolated for improved accuracy. Default is TRUE.}
}
\value{
A matrix with two columns. The x-index of the R-peaks are in the first column and the
        y-values of the ECG signal are in the second column
}
\description{
R-peak detection
}
\seealso{
Other HRV: \code{\link{detect.artifacts.xu}};
  \code{\link{plot.artifacts}};
  \code{\link{qrs.interpolate}}
}

