% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrs_detect.R
\name{qrs_detect}
\alias{qrs_detect}
\title{R-peak detection}
\usage{
qrs_detect(ecg, samplingrate, interpolate = TRUE, invert_polarity = FALSE)
}
\arguments{
\item{ecg}{A vector with the ECG signal}

\item{samplingrate}{The sampling rate of the ECG signal}

\item{interpolate}{Should the position of the R-peak be interpolated for improved accuracy. Default is TRUE.}

\item{invert_polarity}{Should the polarity of the ECG signal be invertedy. Default is FALSE.}
}
\value{
A matrix with two columns. The x-index of the R-peaks are in the first column and the
        y-values of the ECG signal are in the second column
}
\description{
R-peak detection
}
\seealso{
Other HRV: \code{\link{detect_artifacts_xu}},
  \code{\link{plot_artifacts}},
  \code{\link{qrs_interpolate}},
  \code{\link{remove_ibi_artifacts}}
}

