% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrs_detect.R
\name{qrs_interpolate}
\alias{qrs_interpolate}
\title{Improve the identification of the R-peak using cubic spline interpolation}
\usage{
qrs_interpolate(ind.r, sig, window.length = 6, n.spline = 1000)
}
\arguments{
\item{ind.r}{A vector with the indices of the R-peaks}

\item{sig}{A vector containing the ECG signal}

\item{window.length}{The length of the window in samples. Default is 6.}

\item{n.spline}{The number of splines to use.}
}
\value{
A matrix with two columns. The x-index of the R-peaks are in the first column and the
        y-values of the ECG signal are in the second column

http://www.ncbi.nlm.nih.gov/pubmed/9302678
}
\description{
Improve the identification of the R-peak using cubic spline interpolation
}
\references{
Daskalov, I and Christov, I. (1997) Improvement of resolution in measurement of electrocardiogram RR intervals by interpolation. Medical engineering & Physics, 19;4;375-379. Elsevier.
}
\seealso{
Other HRV: \code{\link{detect_artifacts_xu}},
  \code{\link{plot_artifacts}}, \code{\link{qrs_detect}},
  \code{\link{remove_ibi_artifacts}}
}

