% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.block.event.data}
\alias{read.block.event.data}
\title{Read block information from a file or from a database and return
the given blocks as a dataframe.}
\usage{
read.block.event.data(filename, data.format = NULL, casename = NULL,
  data.type = NULL)
}
\arguments{
\item{filename}{The name of the file containing the the block data.}

\item{data.format}{The data format of the blocks or events:

\describe{
\item{csv}{CSV format.}
\item{sqlite, db, db3}{SQLite 3 format.}
\item{xlsx}{Excel XML format.}
}}

\item{casename}{The casename.
#'}

\item{data.type}{The type of data to be read: either \code{block} or \code{event}.}
}
\value{
The blocks or events as a data frame.
}
\description{
Read block information from a file or from a database and return
the given blocks as a dataframe.
}
\seealso{
Other recording: \code{\link{add_block}};
  \code{\link{add_segment_timestamp}};
  \code{\link{block_to_seconds}};
  \code{\link{collect_results_collection}};
  \code{\link{collect_results}};
  \code{\link{create_block_simple}};
  \code{\link{create_block_structure}};
  \code{\link{cut_recording_s}};
  \code{\link{cut_recordings_s}};
  \code{\link{cut_recordings}};
  \code{\link{cut_recording}};
  \code{\link{find_recording_overlap}};
  \code{\link{load_recording}};
  \code{\link{new_recording}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording_get_zerotime}};
  \code{\link{recording_set_blocks}};
  \code{\link{recording_set_events}};
  \code{\link{recording_set_ibi}};
  \code{\link{recording_set_zerotime}};
  \code{\link{save_recording}}
}

