% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.config.excel}
\alias{read.config.excel}
\title{Read the measurement configuration from an Excel file.}
\usage{
read.config.excel(filename)
}
\arguments{
\item{filename}{The Excel file containing the measurement configuration.}
}
\value{
The measurement configuration as a data frame.
}
\description{
Read the measurement configuration from an Excel file.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}
\keyword{internal}

