% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.config}
\alias{read.config}
\title{Read the measurement configuration from a file.}
\usage{
read.config(filename, config.format = NULL)
}
\arguments{
\item{filename}{The file containing the measurement configuration.}

\item{config.format}{Optional. Give the configuration format as a
string. Possible values are
\describe{
\item{sqlite}{SQLite 3 format}
\item{excel}{Excel XML format}
#' }}
}
\value{
The measurement configuration as a data frame.
}
\description{
Read the measurement configuration from a file.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

