% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_edf.R
\name{read.data.edf}
\alias{read.data.edf}
\title{Read data in the European Data Format.}
\usage{
read.data.edf(filename)
}
\arguments{
\item{filename}{The name of the file to be read.}
}
\value{
A recording structure (a list) containing the header information and signal data.
}
\description{
Read data in the European Data Format.
}
\seealso{
Other read_data: \code{\link{read.data.firstbeat}},
  \code{\link{read.data.ibi}},
  \code{\link{read.data.suunto}}, \code{\link{read.data}}
}

