% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_data_firstbeat.R
\name{read.data.firstbeat}
\alias{read.data.firstbeat}
\title{Read data in the Firstbeat Data Format}
\usage{
read.data.firstbeat(filename)
}
\arguments{
\item{filename}{The name of the file to be read.}
}
\value{
A recording structure (a list) containing the header information and signal data.
}
\description{
Read data in the Firstbeat Data Format
}
\seealso{
Other read_data: \code{\link{read.data.edf}};
  \code{\link{read.data.ibi}};
  \code{\link{read.data.suunto}}; \code{\link{read.data}}
}

