% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_ibi.R
\name{read.data.ibi}
\alias{read.data.ibi}
\title{Read data in generic IBI format with one IBI in milliseconds per line.}
\usage{
read.data.ibi(filename)
}
\arguments{
\item{filename}{The name of the file to be read.}
}
\value{
A recording structure (a list) containing the header information and signal data.
}
\description{
Read data in generic IBI format with one IBI in milliseconds per line.
}
\seealso{
Other read_data: \code{\link{read.data.edf}},
  \code{\link{read.data.firstbeat}},
  \code{\link{read.data.suunto}}, \code{\link{read.data}}
}

