% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.data}
\alias{read.data}
\title{Generic function for reading data.}
\usage{
read.data(filename, format = NULL)
}
\arguments{
\item{filename}{The name of the file to be read.}

\item{format}{The format of the data (optional). The supported data formats are:

\describe{
\item{sdf}{Suunto data format.}
\item{fbe}{Firstbeat data format.}
\item{edf}{European Data Format (using the EDF-package).}
}}
}
\value{
A recording structure (a list) containing the header information and signal data.
}
\description{
The input format is determined automatically from the file extension or from the provided \code{format} option.
}
\seealso{
Other read_data: \code{\link{read.data.edf}};
  \code{\link{read.data.firstbeat}};
  \code{\link{read.data.suunto}}
}

