% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.data.suunto}
\alias{read.data.suunto}
\title{Read data in the Suunto Data Format}
\usage{
read.data.suunto(filename)
}
\arguments{
\item{filename}{The name of the file to be read.}
}
\value{
A recording structure (a list) containing the header information and signal data.
}
\description{
Read data in the Suunto Data Format
}
\seealso{
Other read_data: \code{\link{read.data.edf}};
  \code{\link{read.data.firstbeat}};
  \code{\link{read.data.ibi}}; \code{\link{read.data}}
}

