% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.dbtable.sqlite}
\alias{read.dbtable.sqlite}
\title{Read blocks or events from an sqlite database.}
\usage{
read.dbtable.sqlite(filename, dbtable = NULL, casename = NULL,
  query.string = NULL)
}
\arguments{
\item{filename}{The name of the file containing the sqlite database.}

\item{dbtable}{The name of the database table to be raed.}

\item{casename}{The casename.}

\item{query.string}{An optional query string.}
}
\value{
The blocks or events as a data frame.
}
\description{
Read blocks or events from an sqlite database.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}
\keyword{internal}

