% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recording.set.blocks}
\alias{recording.set.blocks}
\title{Store the block information inside the recording structure}
\usage{
recording.set.blocks(recording, filename = NULL, casename = NULL)
}
\arguments{
\item{filename}{The database filename. Optional. If not provided it is determined from the field \code{recording$properties$db.filename}.}

\item{casename}{The casename. Optional. If not provided it is determined from the field \code{recording$properties$casename}.}
}
\value{
The recording structure with the block information added.
}
\description{
Store the block information inside the recording structure
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

