% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recording.set.ibi}
\alias{recording.set.ibi}
\title{Store the interbeat interval data series in the recording structure}
\usage{
recording.set.ibi(recording, ibi, ibi.t = NULL)
}
\arguments{
\item{recording}{The recording structure}

\item{ibi}{An array with the interbeat intervals}

\item{ibi.t}{An optional array with the times of occurrence of the interbeat intervals.}
}
\value{
The recording structure with the interbeat intervals added.
}
\description{
Store the interbeat interval data series in the recording structure
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

