% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recording.set.zerotime}
\alias{recording.set.zerotime}
\title{Set the zerotime of the recording, i.e., the timestamp that corresponds to t = 0.}
\usage{
recording.set.zerotime(recording, timestamp = NULL)
}
\arguments{
\item{timestamp}{The timestamp in ISO-8601 format: YYYYMMDDTHHMMSS. If the timestamp is not given, it is read from the recording.}
}
\value{
The recording structure with the zerotime added.
}
\description{
Given the zerotime as a timestamp, calculate the zerotime in seconds and store both the timestamp
and the number of seconds from the start of the recording when the zerotime occurs.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{save.recording}}
}

