% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\name{recording_set_ibi}
\alias{recording_set_ibi}
\title{Store the interbeat interval data series in the recording structure}
\usage{
recording_set_ibi(recording, ibi, ibi.t = NULL)
}
\arguments{
\item{recording}{The recording structure}

\item{ibi}{An array with the interbeat intervals}

\item{ibi.t}{An optional array with the times of occurrence of the interbeat intervals.}
}
\value{
The recording structure with the interbeat intervals added.
}
\description{
Store the interbeat interval data series in the recording structure
}
\seealso{
Other recording: \code{\link{add_block}},
  \code{\link{add_segment_timestamp}},
  \code{\link{block_to_seconds}},
  \code{\link{collect_results_collection}},
  \code{\link{collect_results}},
  \code{\link{create_block_simple}},
  \code{\link{create_block_structure}},
  \code{\link{cut_recording_s}},
  \code{\link{cut_recordings_s}},
  \code{\link{cut_recordings}},
  \code{\link{cut_recording}},
  \code{\link{find_recording_overlap}},
  \code{\link{load_recording}},
  \code{\link{new_recording}},
  \code{\link{read.block.event.data}},
  \code{\link{read.config.excel}},
  \code{\link{read.config.sqlite}},
  \code{\link{read.config}},
  \code{\link{read.dbtable.excel}},
  \code{\link{read.dbtable.sqlite}},
  \code{\link{recording_get_zerotime}},
  \code{\link{recording_set_blocks}},
  \code{\link{recording_set_events}},
  \code{\link{recording_set_zerotime}},
  \code{\link{save_recording}}
}

