% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recording.R
\name{recording_set_zerotime}
\alias{recording_set_zerotime}
\title{Set the zerotime of the recording, i.e., the timestamp that corresponds to t = 0.}
\usage{
recording_set_zerotime(recording, timestamp = NULL)
}
\arguments{
\item{timestamp}{The timestamp in ISO-8601 format: YYYYMMDDTHHMMSS. If the timestamp is not given, it is read from the recording.}
}
\value{
The recording structure with the zerotime added.
}
\description{
Given the zerotime as a timestamp, calculate the zerotime in seconds and store both the timestamp
and the number of seconds from the start of the recording when the zerotime occurs.
}
\seealso{
Other recording: \code{\link{add_block}};
  \code{\link{add_segment_timestamp}};
  \code{\link{block_to_seconds}};
  \code{\link{collect_results_collection}};
  \code{\link{collect_results}};
  \code{\link{create_block_simple}};
  \code{\link{create_block_structure}};
  \code{\link{cut_recording_s}};
  \code{\link{cut_recordings_s}};
  \code{\link{cut_recordings}};
  \code{\link{cut_recording}};
  \code{\link{find_recording_overlap}};
  \code{\link{load_recording}};
  \code{\link{new_recording}};
  \code{\link{read.block.event.data}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording_get_zerotime}};
  \code{\link{recording_set_blocks}};
  \code{\link{recording_set_events}};
  \code{\link{recording_set_ibi}};
  \code{\link{save_recording}}
}

