% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_ibi_artifacts}
\alias{remove_ibi_artifacts}
\title{Automatically detect and remove artifacts.}
\usage{
remove_ibi_artifacts(recording, signal = "ibi")
}
\arguments{
\item{recording}{A recording structure.}

\item{signal}{A signal name found in the recording. Optional, default is \code{ibi}.}
}
\value{
The recording with the artifacts removed.
}
\description{
Artifact detection is performed using the method presented by Xu
and Schuckers \code{\link{detect_artifacts_xu}}. The interbeat
intervals (IBIs) that are categorized as artifactual are removed
from the IBI series and from the time vector.
}
\seealso{
Other HRV: \code{\link{detect_artifacts_xu}};
  \code{\link{plot_artifacts}}; \code{\link{qrs_detect}};
  \code{\link{qrs_interpolate}}
}

