% Generated by roxygen2 (4.0.2): do not edit by hand
\name{replicate.df.row}
\alias{replicate.df.row}
\title{Replicate a row in a data frame N times returning a new data frame}
\usage{
replicate.df.row(row, N)
}
\arguments{
\item{row}{A data frame row.}

\item{N}{An integer giving the number of times the data frame row should be repeated.}
}
\value{
A data frame with N rows.
}
\description{
Replicate a row in a data frame N times returning a new data frame
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

