% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{replicate_df_row}
\alias{replicate_df_row}
\title{Replicate a row in a data frame N times returning a new data frame}
\usage{
replicate_df_row(row, N)
}
\arguments{
\item{row}{A data frame row.}

\item{N}{An integer giving the number of times the data frame row should be repeated.}
}
\value{
A data frame with N rows.
}
\description{
Replicate a row in a data frame N times returning a new data frame
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{hr_from_ibi}}; \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}};
  \code{\link{timestamp_to_str}}
}

