% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seconds.to.indices}
\alias{seconds.to.indices}
\title{Get the indices corresponding to the times t0 and t1 given in seconds.}
\usage{
seconds.to.indices(recording, t0, t1 = NULL, signal = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{t0}{The start in seconds, from the start of the recording.}

\item{t1}{The stop timestamp (as a string or as a POSIXct timestamp). Optional, default is NULL}

\item{signal}{A signal name found in the recording. Optional, default is NULL.}
}
\value{
The index corresponding to t0, or if t1 is also given, a two-element list with indices corresponding to t0 and t1.
}
\description{
Get the indices corresponding to the times t0 and t1 given in seconds.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{str.to.timestamp}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

