% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seconds_to_indices}
\alias{seconds_to_indices}
\title{Get the indices corresponding to the times t0 and t1 given in seconds.}
\usage{
seconds_to_indices(recording, t0, t1 = NULL, signal = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{t0}{The start in seconds, from the start of the recording.}

\item{t1}{The stop timestamp (as a string or as a POSIXct timestamp). Optional, default is NULL}

\item{signal}{A signal name found in the recording. Optional, default is NULL.}
}
\value{
The index corresponding to t0, or if t1 is also given, a two-element list with indices corresponding to t0 and t1.
}
\description{
Get the indices corresponding to the times t0 and t1 given in seconds.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{hr_from_ibi}}; \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}}
}

