% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_tz_to_utc}
\alias{set_tz_to_utc}
\title{Set the timezone of a given \code{POSIXct} timestamp to UTC.}
\usage{
set_tz_to_utc(ts)
}
\arguments{
\item{ts}{A \code{POSIXct} timestamp}
}
\value{
The timestamp converted to UTC time.
}
\description{
Set the timezone of a given \code{POSIXct} timestamp to UTC.
}

