% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{shade_frequency_band}
\alias{shade_frequency_band}
\title{Shade the area under the power spectrum corresponding to a frequency band.}
\usage{
shade_frequency_band(spec, band, col = "red")
}
\arguments{
\item{spec}{A spect}

\item{band}{A two-element vector with the upper and lower
frequencies of the frequency band to shade.}

\item{col}{The colour to use for shading. Default is \code{red}.}
}
\value{
Nothing. Shades a frequency band.
}
\description{
Shade the area under the power spectrum corresponding to a frequency band.
}
\seealso{
Other visualizations: \code{\link{plot_all_blocks}},
  \code{\link{plot_block}},
  \code{\link{plot_check_rr_detection}},
  \code{\link{plot_ecg_r_peak}}, \code{\link{plot_metric}},
  \code{\link{plot_spectrum}}
}
\keyword{internal}

