% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str.to.timestamp}
\alias{str.to.timestamp}
\title{Convert a string to a timestamp}
\usage{
str.to.timestamp(s, timeformat = "\%Y\%m\%dT\%H\%M\%S")
}
\arguments{
\item{s}{A string containing a timestamp.}

\item{timeformat}{The timeformat used to parse the string. Default is "\%Y\%m\%dT\%H\%M\%S".}
}
\value{
A timestamp.
}
\description{
Convert a string to a timestamp
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{timeintervals.to.seconds}};
  \code{\link{timestamp.to.seconds}}
}

