% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{str_to_timestamp}
\alias{str_to_timestamp}
\title{Convert a string to a timestamp}
\usage{
str_to_timestamp(s, timeformat = "\%Y\%m\%dT\%H\%M\%S")
}
\arguments{
\item{s}{A string containing a timestamp.}

\item{timeformat}{The timeformat used to parse the string. Default is "\%Y\%m\%dT\%H\%M\%S".}
}
\value{
A timestamp.
}
\description{
Convert a string to a timestamp
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{hr_from_ibi}}; \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{num_to_timestamp}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}};
  \code{\link{timestamp_to_str}}
}

