% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timeintervals.to.seconds}
\alias{timeintervals.to.seconds}
\title{Convert a time interval defined by two timestamps to seconds from the start of the recording.}
\usage{
timeintervals.to.seconds(recording, t0, t1, timeformat = NULL,
  return.indices = FALSE, signal = NULL)
}
\arguments{
\item{recording}{A recording structure.}

\item{t0}{The start timestamp (as a string or as a POSIXct timestamp).}

\item{t1}{The stop timestamp (as a string or as a POSIXct timestamp).}

\item{timeformat}{The timeformat used to parse the strings (optional). Default is "\%Y\%m\%dT\%H\%M\%S".}

\item{return.indices}{Boolean indicating whether the indices in the
\code{signal} corresponding to the time intervalshould be
returned. Optional, default is FALSE.}

\item{signal}{A signal name found in the recording. Optional, default is NULL.}
}
\value{
A two-element list with The number of seconds corresponding to the start and stop timestamp, or if
return.indices = TRUE, a four-element list containing the seconds
and the indices corresponding to the time interval.
}
\description{
Convert a time interval defined by two timestamps to seconds from the start of the recording.
}
\seealso{
Other utilities: \code{\link{defactor}};
  \code{\link{ensure.path}};
  \code{\link{extract.segment.block}};
  \code{\link{extract.segment.s}};
  \code{\link{extract.segment.timestamp}};
  \code{\link{extract.segment}};
  \code{\link{generate.result.row}};
  \code{\link{generate.segments.from.block}};
  \code{\link{generate.segments}};
  \code{\link{get.file.extension}};
  \code{\link{get.filename.parts}};
  \code{\link{ibi.from.ecg}};
  \code{\link{integrate.power}};
  \code{\link{replicate.df.row}};
  \code{\link{seconds.to.indices}};
  \code{\link{str.to.timestamp}};
  \code{\link{timestamp.to.seconds}}
}
\keyword{internal}

