% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{timestamp_to_str}
\alias{timestamp_to_str}
\title{Convert a timestamp to a string}
\usage{
timestamp_to_str(ts, timeformat = "\%Y\%m\%dT\%H\%M\%S")
}
\arguments{
\item{ts}{A POSIXct timestamp.}

\item{timeformat}{The timeformat used to format the timestamp as a string. Default is "\%Y\%m\%dT\%H\%M\%S".}
}
\value{
A string.
}
\description{
Convert a timestamp to a string
}
\seealso{
Other utilities: \code{\link{defactor}},
  \code{\link{ensure_path}},
  \code{\link{extract_segment_block}},
  \code{\link{extract_segment_s}},
  \code{\link{extract_segment_timestamp}},
  \code{\link{extract_segment}},
  \code{\link{generate_result_row}},
  \code{\link{generate_segments_from_block}},
  \code{\link{generate_segments}},
  \code{\link{get_file_extension}},
  \code{\link{get_filename_parts}},
  \code{\link{hr_from_ibi}}, \code{\link{ibi_from_ecg}},
  \code{\link{integrate_power}},
  \code{\link{num_to_timestamp}},
  \code{\link{replicate_df_row}},
  \code{\link{seconds_to_indices}},
  \code{\link{str_to_timestamp}},
  \code{\link{timeintervals_to_seconds}},
  \code{\link{timestamp_to_seconds}}
}

