% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{colibri}
\alias{colibri}
\alias{colibri-package}
\title{Platform for automating analyses of physiological time series in R.}
\description{
The colibri package provides functions for automating analysis of
physiological time series in R.
}

