% Generated by roxygen2 (4.0.2): do not edit by hand
\name{defactor}
\alias{defactor}
\title{Convert a factor to a string or numeric representation.}
\usage{
defactor(x, type = "numeric")
}
\arguments{
\item{x}{A factor}

\item{type}{A string indicating the type of the factor: "numeric" (default) or "string).}
}
\value{
A string representation of the factor.
}
\description{
Convert a factor to a string or numeric representation.
}
\seealso{
Other utilities: \code{\link{ensure_path}};
  \code{\link{extract_segment_block}};
  \code{\link{extract_segment_s}};
  \code{\link{extract_segment_timestamp}};
  \code{\link{extract_segment}};
  \code{\link{generate_result_row}};
  \code{\link{generate_segments_from_block}};
  \code{\link{generate_segments}};
  \code{\link{get_file_extension}};
  \code{\link{get_filename_parts}};
  \code{\link{ibi_from_ecg}};
  \code{\link{integrate_power}};
  \code{\link{replicate_df_row}};
  \code{\link{seconds_to_indices}};
  \code{\link{str_to_timestamp}};
  \code{\link{timeintervals_to_seconds}};
  \code{\link{timestamp_to_seconds}}
}

