% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lombscargle}
\alias{lombscargle}
\title{Lomb-Scargle Periodogram estimation}
\usage{
lombscargle(t, x, f = NULL, demean = TRUE, smooth = FALSE,
  smooth.kernel = "daniell", smooth.degree = 5,
  normalization = "variance")
}
\arguments{
\item{t}{A vector with the time instants of the events.}

\item{x}{A vector with}

\item{f}{A vector of frequencies at which the spectrum should be esimated.}

\item{demean}{Should the mean be subtraced from the spectrum. Default is TRUE.}

\item{smooth}{Should the spectrum be smoothed. Default is FALSE.}

\item{smooth.kernel}{The smoothing kernel. The default is to use a 'daniell' kernel.}

\item{smooth.degree}{The degree of smoothing. Default is 5.}

\item{normalization}{How to normalize the spectrum. Possible values are 'variance' (default) and 'none'.}
}
\value{
A list with the following components.
\describe{
\item{f}{The frequencies at which the spectrum was estimated.}
\item{Px}{The power at the frequencies in f.}
\item{Prob}{The probability of the power in a particular frequency band}
}
}
\description{
This code is adapted from the function lomb.m from
http://www.mit.edu/~gari/CODE/HRV/CliffordHRVtools.zip.
}
\seealso{
http://www.ltrr.arizona.edu/~dmeko/notes_6.pdf
https://onlinecourses.science.psu.edu/stat510/?q=book/export/html/57
http://reference.wolfram.com/applications/timeseries/UsersGuideToTimeSeries/SpectralAnalysis/1.8.4.html

Other HRV frequency-domain: \code{\link{analyse.frequencydomain.helper}};
  \code{\link{analyse.frequencydomain}};
  \code{\link{ibi.band.custom}};
  \code{\link{ibi.band.standard}}
}

