% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.block.event.data}
\alias{read.block.event.data}
\title{Read block information from a file or from a database and return
the given blocks as a dataframe.}
\usage{
read.block.event.data(filename, data.format = NULL, casename = NULL,
  data.type = NULL)
}
\arguments{
\item{filename}{The name of the file containing the the block data.}

\item{data.format}{The data format of the blocks or events:

\describe{
\item{csv}{CSV format.}
\item{sqlite, db, db3}{SQLite 3 format.}
\item{xlsx}{Excel XML format.}
}}

\item{casename}{The casename.
#'}

\item{data.type}{The type of data to be read: either \code{block} or \code{event}.}
}
\value{
The blocks or events as a data frame.
}
\description{
Read block information from a file or from a database and return
the given blocks as a dataframe.
}
\seealso{
Other recording: \code{\link{add.block}};
  \code{\link{add.segment.timestamp}};
  \code{\link{block.to.seconds}};
  \code{\link{collect.results.collection}};
  \code{\link{collect.results}};
  \code{\link{create.block.simple}};
  \code{\link{create.block.structure}};
  \code{\link{cut.recording.s}};
  \code{\link{cut.recordings.s}};
  \code{\link{cut.recordings}};
  \code{\link{cut.recording}};
  \code{\link{find.recording.overlap}};
  \code{\link{load.recording}};
  \code{\link{new.recording}};
  \code{\link{read.config.excel}};
  \code{\link{read.config.sqlite}};
  \code{\link{read.config}};
  \code{\link{read.dbtable.excel}};
  \code{\link{read.dbtable.sqlite}};
  \code{\link{recording.get.zerotime}};
  \code{\link{recording.set.blocks}};
  \code{\link{recording.set.events}};
  \code{\link{recording.set.ibi}};
  \code{\link{recording.set.zerotime}};
  \code{\link{save.recording}}
}

