% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_correction.R
\name{remove_ibi_artifacts}
\alias{remove_ibi_artifacts}
\title{Automatically detect and remove artifacts.}
\usage{
remove_ibi_artifacts(recording, signal = "ibi", signal2 = "ibi.amp")
}
\arguments{
\item{recording}{A recording structure.}

\item{signal}{A signal name found in the recording. Optional, default is \code{ibi}.}

\item{signal2}{An additional signal from which the indices found to be artifacts in \code{signal} are removed too. Default is \code{ibi.amp}.}
}
\value{
The recording with the artifacts removed.
}
\description{
Artifact detection is performed using the method presented by Xu
and Schuckers \code{\link{detect_artifacts_xu}}. The interbeat
intervals (IBIs) that are categorized as artifactual are removed
from the IBI series and from the time vector.
}
\seealso{
Other HRV: \code{\link{detect_artifacts_xu}},
  \code{\link{plot_artifacts}}, \code{\link{qrs_detect}},
  \code{\link{qrs_interpolate}}
}

